02-use-cases/customer-support-assistant/requirements.txt
strands-agents
strands-agents-tools
requests
PyYAML
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
fastmcp
pydantic
bedrock-agentcore
bedrock-agentcore-starter-toolkit
botocore
boto3
playwright
nest-asyncio